Dynamically update the greeter text
Adds new flag: `--greeter-cmd`, which accepts a command name to execute
and sets the output to the greeter text. If the screen is idle, the
command gets executed every second, if the user is typing the command
gets executed in every keypress.

This new flag conflicts with `--greeter-text` because when the output
from the command gets captured it overrides the variable containing the
greeter text passed with `--greeter-text`, so they can't be used
together.